requirements.txt
apache-airflow==3.0.6
pandas==2.3.2
requests==2.32.5